inventory/pyproject.toml
[tool.poetry]
name = "inventory"
version = "0.1.0"
description = ""
authors = ["Aneeq Khatri <aneeqabdulsamad5761@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
requests = "^2.31.0"
uvicorn = "^0.28.0"
httpx = "^0.27.0"
sqlmodel = "^0.0.16"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.9"
pytest = "^8.1.1"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
